feat: refactor linkedin api wrapper
