+++ linkedin/__init__.py
@@ -1,0 +1,7 @@
+"""Python Linkedin API Wrapper.
+
+This is a python package that allows for simple access to
+Linkedin's API. We aim to be as easy to use as possible.
+"""
+
+from .api import __version__  # NOQA
