+++ linkedin/api/constants.py
@@ -1,0 +1,6 @@
+"""Linkedin constants."""
+from .endpoints import API_PATH  # noqa
+
+__version__ = "0.1.0"
+
+USER_AGENT_FORMAT = f"LikedIn/{__version__}"
