+++ linkedin/api/endpoints.py
@@ -1,0 +1,8 @@
+"""List of API endpoints Linkedin knows about."""
+
+# flake8: noqa
+# fmt: off
+API_PATH = {
+    "me":            "me/",
+    "reactions":     "reactions?actor={params}",
+}
