+++ linkedin/api/service/auth.py
@@ -1,0 +1,8 @@
+"""Provide `/authorization` service class."""
+from .base import ServiceBase
+
+
+class Auth(ServiceBase):
+    """Auth is a Service class that represents the `/authorization` endpoint."""
+
+    ...
