+++ linkedin/api/service/reactions.py
@@ -1,0 +1,8 @@
+"""Provide `/reactions` service class."""
+from .base import ServiceBase
+
+
+class Reactions(ServiceBase):
+    """Reactions is a Service class that represents the `/reactions` endpoint."""
+
+    ...
